flask_server/requirements.txt
# Flask core
Flask==3.0.0
Flask-CORS==4.0.0

# MongoDB
pymongo==4.6.0

# Environment management
python-dotenv==1.0.0

# HTTP requests (for downloading video from Cloudinary)
requests>=2.32.5

# Video processing (for converting webm to mp4)
moviepy==1.0.3

# Pathway (for document parsing - version 0.27.0 includes PaddleOCRParser)
# Note: pathway[xpack-llm-docs] installs required dependencies including newer paddleocr with PPStructureV3
pathway==0.27.0
pathway[xpack-llm-docs]

# Cloudinary for document storage
cloudinary==1.41.0

# Password hashing
bcrypt==4.1.2

# MCP client for SmartFolio XAI
fastmcp>=0.1.0
